+++ lib/ecto/migration/schema_migration.ex
@@ -14,8 +14,8 @@ defmodule Ecto.Migration.SchemaMigration do
 
   @opts [timeout: :infinity, log: false]
 
-  def ensure_schema_migrations_table!(repo, opts) do
-    {repo, source} = get_repo_and_source(repo)
+  def ensure_schema_migrations_table!(repo, config, opts) do
+    {repo, source} = get_repo_and_source(repo, config)
     table_name = String.to_atom(source)
     table = %Ecto.Migration.Table{name: table_name, prefix: opts[:prefix]}
     meta = Ecto.Adapter.lookup_meta(repo.get_dynamic_repo())
@@ -29,31 +29,27 @@ defmodule Ecto.Migration.SchemaMigration do
     repo.__adapter__().execute_ddl(meta, {:create_if_not_exists, table, commands}, @opts)
   end
 
-  def versions(repo, prefix) do
-    {_repo, source} = get_repo_and_source(repo)
-
-    from(m in source, select: type(m.version, :integer))
-    |> Map.put(:prefix, prefix)
+  def versions(repo, config) do
+    {repo, source} = get_repo_and_source(repo, config)
+    {repo, from(m in source, select: type(m.version, :integer))}
   end
 
-  def up(repo, version, prefix) do
-    {repo, source} = get_repo_and_source(repo)
+  def up(repo, config, version, prefix) do
+    {repo, source} = get_repo_and_source(repo, config)
 
     %__MODULE__{version: version}
     |> Ecto.put_meta(source: source)
     |> repo.insert([prefix: prefix] ++ @opts)
   end
 
-  def down(repo, version, prefix) do
-    {repo, source} = get_repo_and_source(repo)
+  def down(repo, config, version, prefix) do
+    {repo, source} = get_repo_and_source(repo, config)
 
     from(m in source, where: m.version == type(^version, :integer))
     |> repo.delete_all([prefix: prefix] ++ @opts)
   end
 
-  def get_repo_and_source(repo) do
-    config = repo.config()
-
+  def get_repo_and_source(repo, config) do
     {Keyword.get(config, :migration_repo, repo),
      Keyword.get(config, :migration_source, "schema_migrations")}
   end